Page for new Coop Details mostly done
